feat: replate create app use
