+++ CreateAppModal/interface.ts
@@ -1,7 +1,2 @@
-export enum CREATE_TYPE {
-  FROM_RESOURCE = "FROM_RESOURCE",
-  FROM_TEMPLATE = "FROM_TEMPLATE",
-}
-
 export * from "./CreateFromResource/interface"
 export * from "./CreateFromTemplate/interface"
